chapter 7

+++ ch06/src/Ex04_CallMethod.java
@@ -1,9 +1,8 @@
-    // 메서드의 호출
 public class Ex04_CallMethod {
-
+    // 메서드의 호출
     public static void main(String[] args) {
         MethodCall mc = new MethodCall();
-        long result = mc.add(5L, 3L);
+        double result = mc.add(5L, 3L);
 
         System.out.println("add(5L, 3l) = " + result);
     }
